1)Fix for recurrent message which was stopping at first run if number of occurences and end date were not provided 2)Integration tests for recurrent message domain and service classes

+++ grails-app/services/net/hedtech/banner/general/communication/recurrence/CommunicationRecurrentMessageCompositeService.groovy
@@ -180,7 +180,7 @@ class CommunicationRecurrentMessageCompositeService {
         recurrentMessage = (CommunicationRecurrentMessage) communicationRecurrentMessageService.update(recurrentMessage)
 
         //if no of occurences <= total count, delete the cron_trigger entry
-        if(recurrentMessage.noOfOccurrences <= recurrentMessage.totalCount) {
+        if(recurrentMessage.noOfOccurrences && recurrentMessage.noOfOccurrences <= recurrentMessage.totalCount) {
             if( recurrentMessage.jobId != null ) {
                 schedulerJobService.deleteScheduledJob( recurrentMessage.jobId, recurrentMessage.groupId )
             }
